Make sure we do not reimport singletons. Should fix #379
